library skeleton
