+++ .gitignore
@@ -117,9 +117,9 @@ dmypy.json
 .idea
 
 # Simplex
-experiments/results/*
-experiments/data/*
-experiments/time_series_toy.py
+src/simplexai/experiments/results/*
+src/simplexai/experiments/data/*
+src/simplexai/experiments/time_series_toy.py
 data/*
 notebooks/*
 !notebooks/*.ipynb
