+++ .isort.cfg
@@ -1,0 +1,7 @@
+[settings]
+profile = black
+multi_line_output = 3
+include_trailing_comma = True
+force_grid_wrap = 0
+use_parentheses = True
+line_length = 88
