+++ .pre-commit-config.yaml
@@ -1,0 +1,38 @@
+exclude: 'setup.py|^docs|^experiments|^tests'
+
+repos:
+- repo: https://github.com/pre-commit/pre-commit-hooks
+  rev: v3.4.0
+  hooks:
+  - id: trailing-whitespace
+  - id: check-added-large-files
+  - id: check-ast
+  - id: check-json
+  - id: check-merge-conflict
+  - id: check-xml
+  - id: check-yaml
+  - id: debug-statements
+  - id: check-executables-have-shebangs
+  - id: end-of-file-fixer
+  - id: requirements-txt-fixer
+  - id: mixed-line-ending
+    args: ['--fix=auto']  # replace 'auto' with 'lf' to enforce Linux/Mac line endings or 'crlf' for Windows
+
+- repo: https://github.com/pycqa/isort
+  rev: 5.8.0
+  hooks:
+  - id: isort
+
+- repo: https://github.com/psf/black
+  rev: 22.3.0
+  hooks:
+  - id: black
+    language_version: python3
+- repo: https://gitlab.com/pycqa/flake8
+  rev: 3.9.1
+  hooks:
+  - id: flake8
+    args: [
+        "--max-line-length=400",
+        "--extend-ignore=E203,W503"
+    ]
